refactor specialties and hero default characteristics
