+++ src/test/java/com/cleancodeheroes/hero/domain/SpecialtyTest.java
@@ -7,13 +7,13 @@ public class SpecialtyTest {
 
     @Test
     public void shouldCreateTank () {
-        var specialty = Specialty.Tank() ;
+        var specialty = SpecialtyCharacteristics.Tank() ;
         Assertions.assertEquals(CharacterType.Tank, specialty.getType());
     }
 
     @Test
     public void shouldValidateTankBaseStats () {
-        var tank = Specialty.Tank() ;
+        var tank = SpecialtyCharacteristics.Tank() ;
         Assertions.assertEquals(1000, tank.getBaseHealthPoints());
         Assertions.assertEquals(100, tank.getBasePower());
         Assertions.assertEquals(20, tank.getBaseArmour());
@@ -21,20 +21,20 @@ public class SpecialtyTest {
 
     @Test
     public void shouldValidateTankDammageMap () {
-        var tank = Specialty.Tank() ;
+        var tank = SpecialtyCharacteristics.Tank() ;
         Assertions.assertTrue(tank.getAdditionalDammage().containsKey(CharacterType.Sorcerer));
         Assertions.assertEquals(20, tank.getAdditionalDammage().get(CharacterType.Sorcerer));
     }
 
     @Test
     public void shouldCreateAssassin () {
-        var specialty = Specialty.Assasin() ;
+        var specialty = SpecialtyCharacteristics.Assasin() ;
         Assertions.assertEquals(CharacterType.Assassin, specialty.getType());
     }
 
     @Test
     public void shouldValidateAssassinBaseStats () {
-        var assasin = Specialty.Assasin() ;
+        var assasin = SpecialtyCharacteristics.Assasin() ;
         Assertions.assertEquals(800, assasin.getBaseHealthPoints());
         Assertions.assertEquals(200, assasin.getBasePower());
         Assertions.assertEquals(5, assasin.getBaseArmour());
@@ -42,20 +42,20 @@ public class SpecialtyTest {
 
     @Test
     public void shouldValidateAssassinDammageMap () {
-        var assassin = Specialty.Assasin() ;
+        var assassin = SpecialtyCharacteristics.Assasin() ;
         Assertions.assertTrue(assassin.getAdditionalDammage().containsKey(CharacterType.Tank));
         Assertions.assertEquals(30, assassin.getAdditionalDammage().get(CharacterType.Tank));
     }
 
     @Test
     public void shouldCreateSorcerer () {
-        var specialty = Specialty.Sorcerer() ;
+        var specialty = SpecialtyCharacteristics.Sorcerer() ;
         Assertions.assertEquals(CharacterType.Sorcerer, specialty.getType());
     }
 
     @Test
     public void shouldValidateSorcererBaseStats () {
-        var sorcerer = Specialty.Sorcerer() ;
+        var sorcerer = SpecialtyCharacteristics.Sorcerer() ;
         Assertions.assertEquals(700, sorcerer.getBaseHealthPoints());
         Assertions.assertEquals(150, sorcerer.getBasePower());
         Assertions.assertEquals(10, sorcerer.getBaseArmour());
@@ -63,7 +63,7 @@ public class SpecialtyTest {
 
     @Test
     public void shouldValidateSorcererDammageMap () {
-        var sorcerer = Specialty.Sorcerer() ;
+        var sorcerer = SpecialtyCharacteristics.Sorcerer() ;
         Assertions.assertTrue(sorcerer.getAdditionalDammage().containsKey(CharacterType.Assassin));
         Assertions.assertEquals(25, sorcerer.getAdditionalDammage().get(CharacterType.Assassin));
     }
